more tags and some syntax

+++ src/ariadne_tringls.h
@@ -33,9 +33,9 @@ extern "C" {
 
 struct Point;
 
-struct Triangle { // sizeof = 16
+struct Triangle {
   short points[3];
-  short tags[3]; // Id of each triangle neighbour of this one
+  int tags[3]; // Id of each triangle neighbour of this one
   unsigned char tree_alt; // 255 is a special value here
   unsigned char field_D;
   unsigned short field_E;
